[@tomclement] Fix failing functional tests on CI
Functional tests were failing on CI because permissions were not being
set up. This commit should fix that. However in the long run we need to
ensure that we run all our tests (not just functional tests) on CI in
the same way we run them locally. Currently this is not the case.